+++ ios/Classes/FlMlKitTextRecognizeMethodCall.swift
@@ -9,8 +9,12 @@ import MLKitVision
 
 class FlMlKitTextRecognizeMethodCall: FlCameraMethodCall {
     var options: CommonTextRecognizerOptions = TextRecognizerOptions()
-    var analyzing: Bool = false
-    var scan: Bool = false
+
+    private var canScan: Bool = false
+
+    private var frequency: Double = 0
+
+    private var lastCurrentTime: TimeInterval = 0
 
     override init(_ _registrar: FlutterPluginRegistrar) {
         super.init(_registrar)
@@ -19,11 +23,13 @@ class FlMlKitTextRecognizeMethodCall: FlCameraMethodCall {
     override func handle(call: FlutterMethodCall, result: @escaping FlutterResult) {
         switch call.method {
         case "startPreview":
+            frequency = (call.arguments as! [String: Any?])["frequency"] as! Double
             startPreview({ [self] sampleBuffer in
-                if !analyzing, scan {
-                    analyzing = true
+                let currentTime = Date().timeIntervalSince1970 * 1000
+                if currentTime - lastCurrentTime >= frequency, canScan {
                     let buffer = CMSampleBufferGetImageBuffer(sampleBuffer)
                     self.analysis(buffer!.image, nil)
+                    lastCurrentTime = currentTime
                 }
             }, call: call, result)
         case "setRecognizedLanguage":
@@ -41,13 +47,8 @@ class FlMlKitTextRecognizeMethodCall: FlCameraMethodCall {
                 }
             }
             result([])
-        case "getScanState":
-            result(scan)
         case "scan":
-            let argument = call.arguments as! Bool
-            if argument != scan {
-                scan = argument
-            }
+            canScan = call.arguments as! Bool
             result(true)
         default:
             super.handle(call: call, result: result)
@@ -87,7 +88,6 @@ class FlMlKitTextRecognizeMethodCall: FlCameraMethodCall {
                     result!(map)
                 }
             }
-            analyzing = false
         }
     }
 }