+++ lib/src/mlkit_text_recognize.dart
@@ -1,26 +1,30 @@
 part of '../fl_mlkit_text_recognize.dart';
 
-typedef EventBarcodeListen = void Function(AnalysisTextModel text);
+typedef EventRecognizedListen = void Function(AnalysisTextModel text);
+typedef FlMlKitTextRecognizeCreateCallback = void Function(
+    FlMlKitTextRecognizeController controller);
 
 class FlMlKitTextRecognize extends StatefulWidget {
   const FlMlKitTextRecognize({
     Key? key,
-    this.onListen,
+    this.onDataChanged,
     this.overlay,
     this.uninitialized,
-    this.onFlashChange,
+    this.onFlashChanged,
     this.autoScanning = true,
-    this.onZoomChange,
+    this.onZoomChanged,
     this.camera,
     this.resolution = CameraResolution.high,
     this.updateReset = false,
     this.fit = BoxFit.fitWidth,
     this.recognizedLanguage = RecognizedLanguage.latin,
+    this.onCreateView,
+    this.frequency = 500,
   }) : super(key: key);
 
   /// 识别回调
   /// Identify callback
-  final EventBarcodeListen? onListen;
+  final EventRecognizedListen? onDataChanged;
 
   /// 显示在预览框上面
   /// Display above preview box
@@ -31,11 +35,11 @@ class FlMlKitTextRecognize extends StatefulWidget {
   final Widget? uninitialized;
 
   /// Flash change
-  final ValueChanged<FlashState>? onFlashChange;
+  final ValueChanged<FlashState>? onFlashChanged;
 
   /// 缩放变化
   /// zoom ratio
-  final ValueChanged<CameraZoomState>? onZoomChange;
+  final ValueChanged<CameraZoomState>? onZoomChanged;
 
   /// 更新组件时是否重置相机
   /// Reset camera when updating components
@@ -60,90 +64,97 @@ class FlMlKitTextRecognize extends StatefulWidget {
   /// Language to be recognized
   final RecognizedLanguage recognizedLanguage;
 
+  /// 如果你设置为10, 2次解析数据间隔为10 毫秒，数字越小 在ios上cpu占有率越高，数字越大，识别速度会变慢，建议设置500-1500
+  /// If you set it to 10, The interval between data parsing is 10 milliseconds
+  /// The larger the number, the slower the parsing,If the number is too small, the CPU percentage will be too high on ios
+  /// Therefore, the recommended setting range is 500 to 1500
+  final double frequency;
+
+  /// get Controller
+  final FlMlKitTextRecognizeCreateCallback? onCreateView;
+
   @override
   _FlMlKitTextRecognizeState createState() => _FlMlKitTextRecognizeState();
 }
 
 class _FlMlKitTextRecognizeState extends FlCameraState<FlMlKitTextRecognize> {
+  FlashState? _flashState;
+  CameraZoomState? _zoomState;
+
   @override
   void initState() {
-    currentChannel = _flMlKitTextRecognizeChannel;
+    controller = FlMlKitTextRecognizeController();
     super.initState();
     uninitialized = widget.uninitialized;
-    WidgetsBinding.instance!.addPostFrameCallback((Duration time) => init());
-  }
-
-  Future<void> init() async {
-    boxFit = widget.fit;
-
-    /// Add message callback
-    await initEvent(eventListen);
-
-    await FlMlKitTextRecognizeMethodCall()
-        .setRecognizedLanguage(widget.recognizedLanguage);
-
-    /// Initialize camera
-    initCamera(camera: widget.camera, resolution: widget.resolution)
-        .then((bool value) {
-      if (!value) return;
-      if (mounted) {
-        setState(() {});
-
-        /// Start scan
-        if (widget.autoScanning) FlMlKitTextRecognizeMethodCall().start();
+    WidgetsBinding.instance!.addPostFrameCallback((Duration time) async {
+      await controller.initialize();
+      if (widget.onCreateView != null) {
+        widget.onCreateView!(controller as FlMlKitTextRecognizeController);
       }
+      initialize();
     });
+    controller.addListener(changedListener);
   }
 
-  void eventListen(dynamic data) {
-    if (widget.onListen != null) {
-      if (data is Map) {
-        widget.onListen!(AnalysisTextModel.fromMap(data));
+  Future<void> initialize() async {
+    var camera = widget.camera;
+    if (camera == null) {
+      final List<CameraInfo>? cameras = controller.cameras;
+      if (cameras == null) return;
+      for (final CameraInfo cameraInfo in cameras) {
+        if (cameraInfo.lensFacing == CameraLensFacing.back) {
+          camera = cameraInfo;
+          break;
+        }
+      }
+    }
+    if (camera == null) return;
+    var textController = controller as FlMlKitTextRecognizeController;
+    final data = await textController.initialize();
+    if (data) {
+      await textController.setRecognizedLanguage(widget.recognizedLanguage);
+      if (widget.onDataChanged != null) {
+        textController.onDataChanged = widget.onDataChanged;
+      }
+      final options = await textController.startPreview(camera,
+          resolution: widget.resolution, frequency: widget.frequency);
+      if (options != null && mounted) {
+        if (widget.autoScanning) textController.startScan();
+        setState(() {});
       }
     }
   }
 
-  @override
-  void onZoomChange(CameraZoomState state) {
-    super.onZoomChange(state);
-    if (widget.onZoomChange != null) widget.onZoomChange!(state);
-  }
-
-  @override
-  void onFlashChange(FlashState state) {
-    super.onFlashChange(state);
-    if (widget.onFlashChange != null) widget.onFlashChange!(state);
+  void changedListener() {
+    if (widget.onFlashChanged != null &&
+        controller.cameraFlash != null &&
+        controller.cameraFlash != _flashState) {
+      _flashState = controller.cameraFlash!;
+      widget.onFlashChanged!(_flashState!);
+    }
+    if (widget.onZoomChanged != null &&
+        controller.cameraZoom != null &&
+        controller.cameraZoom != _zoomState) {
+      _zoomState = controller.cameraZoom!;
+      widget.onZoomChanged!(_zoomState!);
+    }
   }
 
   @override
   void didUpdateWidget(covariant FlMlKitTextRecognize oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (oldWidget.overlay != widget.overlay ||
-        oldWidget.onFlashChange != widget.onFlashChange ||
-        oldWidget.uninitialized != widget.uninitialized ||
-        oldWidget.autoScanning != widget.autoScanning ||
-        oldWidget.fit != widget.fit ||
-        oldWidget.onListen != widget.onListen) {
-      uninitialized = widget.uninitialized;
-      if (widget.updateReset) {
-        cameraMethodCall.dispose().then((bool value) {
-          if (value) init();
-        });
-      }
-    }
-  }
-
-  @override
-  void didChangeAppLifecycleState(AppLifecycleState state) {
-    if (state == AppLifecycleState.resumed) {
-      init();
-    } else {
-      super.didChangeAppLifecycleState(state);
+        oldWidget.onFlashChanged != widget.onFlashChanged ||
+        oldWidget.onZoomChanged != widget.onZoomChanged ||
+        oldWidget.camera != widget.camera ||
+        oldWidget.autoScanning != widget.autoScanning) {
+      if (widget.updateReset) controller.resetCamera();
     }
   }
 
   @override
   Widget build(BuildContext context) {
+    boxFit = widget.fit;
     Widget camera = super.build(context);
     if (widget.overlay != null) {
       camera = Stack(children: <Widget>[
@@ -157,6 +168,7 @@ class _FlMlKitTextRecognizeState extends FlCameraState<FlMlKitTextRecognize> {
   @override
   void dispose() {
     super.dispose();
-    FlMlKitTextRecognizeMethodCall().pause();
+    controller.removeListener(changedListener);
+    controller.dispose();
   }
 }