+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.0
+ * update [`fl_camera`](https://pub.dev/packages/fl_camera)
+ * add intercept callback
+ * remove `FlMlKitTextRecognizeMethodCall()`
+ * add `onCreateView` for `FlMlKitTextRecognizeController()`
+ * add `frequency` for `FlMlKitTextRecognizeController()`
+ * add `stopPreview()` for `FlMlKitTextRecognizeController()`
+ * add `switchCamera()` for `FlMlKitTextRecognizeController()`
+ * add `resetCamera()` for `FlMlKitTextRecognizeController()`
+ * modify `onFlashChange` to `onFlashChanged`
+ * modify `onZoomChange` to `onZoomChanged`
+ * modify `onListen` to `onDataChanged`
 ## 0.1.2
  * update [`fl_camera`](https://pub.dev/packages/fl_camera)
 ## 0.1.1